+++ packages/react/runtime/__test__/renderToOpcodes.test.jsx
@@ -4,21 +4,28 @@ import { Component, createContext, Fragment } from 'preact';
 import { useMemo, useState } from 'preact/hooks';
 import { afterEach, beforeAll, beforeEach, describe, expect, it, vi } from 'vitest';
 
-import { elementTree } from './utils/nativeMethod';
+import { elementTree, waitSchedule } from './utils/nativeMethod';
 import { globalEnvManager } from './utils/envManager';
 import { setupDocument } from '../src/document';
 import { renderOpcodesInto } from '../src/opcodes';
 import renderToString from '../src/renderToOpcodes';
 import { setupPage, SnapshotInstance, snapshotInstanceManager } from '../src/snapshot';
 import { createElement, cloneElement } from '../lepus';
+import { Suspense } from 'preact/compat';
+import { createSuspender } from './createSuspender';
 
 describe('renderToOpcodes', () => {
   beforeAll(() => {
     globalEnvManager.switchToMainThread();
   });
 
+  afterEach(() => {
+    vi.clearAllMocks();
+  });
+
   it('should render hello world', () => {
-    expect(renderToString(function() {})).toMatchInlineSnapshot(`[]`);
+    expect(renderToString(function() {
+    })).toMatchInlineSnapshot(`[]`);
 
     expect(
       renderToString(
@@ -57,6 +64,7 @@ describe('renderToOpcodes', () => {
       }, []);
       return <view>{a}</view>;
     }
+
     expect(
       renderToString(
         <view>
@@ -95,10 +103,12 @@ describe('renderToOpcodes', () => {
       static getDerivedStateFromProps(props, state) {
         return { a: 1 };
       }
+
       render() {
         return <view>{this.state.a}</view>;
       }
     }
+
     expect(
       renderToString(
         <view>
@@ -141,6 +151,7 @@ describe('renderToOpcodes', () => {
         </view>
       );
     }
+
     expect(renderToString(<App />)).toMatchInlineSnapshot(`
       [
         0,
@@ -170,6 +181,7 @@ describe('renderToOpcodes', () => {
         </view>
       );
     }
+
     expect(renderToString(<App />)).toMatchInlineSnapshot(`
       [
         0,
@@ -316,6 +328,7 @@ describe('renderToOpcodes', () => {
     function App() {
       undefined();
     }
+
     expect(
       () =>
         renderToString(
@@ -333,8 +346,10 @@ describe('renderToOpcodes', () => {
 
   it('should throw when error occur - with ErrorBoundary ignored', () => {
     const f = vi.fn();
+
     class ErrorBoundary extends Component {
       componentDidCatch = f;
+
       render() {
         return this.props.children;
       }
@@ -343,6 +358,7 @@ describe('renderToOpcodes', () => {
     function App() {
       undefined();
     }
+
     expect(
       () =>
         renderToString(
@@ -358,6 +374,7 @@ describe('renderToOpcodes', () => {
 
     class ErrorBoundary2 extends Component {
       static getDerivedStateFromError = f;
+
       render() {
         return this.props.children;
       }
@@ -379,6 +396,200 @@ describe('renderToOpcodes', () => {
     // renderToString will throw on Error without calling `options[DIFFED]`
     vi.mocked(console.profile).mockClear();
     vi.mocked(console.profileEnd).mockClear();
+  });
+
+  it('should render fallback when a direct child suspends', async () => {
+    const { Suspender, suspended } = createSuspender();
+
+    const fallbackJsx = <text>loading...</text>;
+
+    const rendered = renderToString(
+      <Suspense fallback={fallbackJsx}>
+        <Suspender>
+          <text className='foo'>bar</text>
+        </Suspender>
+      </Suspense>,
+    );
+
+    expect(rendered.length).toBe(3);
+    expect(rendered[0]).toStrictEqual(0);
+    expect(rendered[1].type).toStrictEqual(fallbackJsx.type);
+    expect(rendered[2]).toStrictEqual(1);
+  });
+
+  it('should render fallback when suspended component is not a direct child', async () => {
+    const { Suspender, suspended } = createSuspender();
+
+    const fallbackJsx = <text>loading...</text>;
+
+    const rendered = renderToString(
+      <Suspense fallback={fallbackJsx}>
+        <view attr={Math.random()}>
+          <Suspender>
+            <text className='foo'>bar</text>
+          </Suspender>
+        </view>
+      </Suspense>,
+    );
+
+    expect(rendered.length).toBe(3);
+    expect(rendered[0]).toStrictEqual(0);
+    expect(rendered[1].type).toStrictEqual(fallbackJsx.type);
+    expect(rendered[2]).toStrictEqual(1);
+  });
+
+  it('should render a fallback that is a Fragment', async () => {
+    const { Suspender, suspended } = createSuspender();
+
+    const fallbackJsx1 = <text>{`loading1...`}</text>;
+    const fallbackJsx2 = <text>{`loading2...`}</text>;
+
+    const fallbackJsx = (
+      <>
+        {fallbackJsx1}
+        {fallbackJsx2}
+      </>
+    );
+
+    const rendered = renderToString(
+      <Suspense fallback={fallbackJsx}>
+        <view attr={Math.random()}>
+          <Suspender>
+            <text className='foo'>bar</text>
+          </Suspender>
+        </view>
+      </Suspense>,
+    );
+
+    expect(rendered[1].type).toStrictEqual(fallbackJsx1.type);
+    expect(rendered[6].type).toStrictEqual(fallbackJsx2.type);
+  });
+
+  it('should render outer fallback when nested child suspends', async () => {
+    const { Suspender: Suspender1, suspended: suspended1 } = createSuspender();
+    const { Suspender: Suspender2, suspended: suspended2 } = createSuspender();
+
+    const fallbackJsx1 = <text>loading1...</text>;
+    const fallbackJsx2 = <text>loading2...</text>;
+
+    const rendered = renderToString(
+      <Suspense fallback={fallbackJsx1}>
+        <view attr={Math.random()}>
+          <Suspender1>
+            <text className='foo'>bar</text>
+            <Suspense fallback={fallbackJsx2}>
+              <Suspender2>
+                <text className='foo'>bar</text>
+              </Suspender2>
+            </Suspense>
+          </Suspender1>
+        </view>
+      </Suspense>,
+    );
+
+    expect(rendered.length).toBe(3);
+    expect(rendered[0]).toStrictEqual(0);
+    expect(rendered[1].type).toStrictEqual(fallbackJsx1.type);
+    expect(rendered[2]).toStrictEqual(1);
+  });
+
+  it('should render inner fallback and resolved content when outer suspense is resolved', async () => {
+    const { Suspender: Suspender1, suspended: suspended1 } = createSuspender();
+    const { Suspender: Suspender2, suspended: suspended2 } = createSuspender();
+    suspended1.resolve();
+    await waitSchedule();
+
+    const fallbackJsx1 = <text>loading1...</text>;
+    const fallbackJsx2 = <text>loading2...</text>;
+    const resolvedJsx1 = <text className='foo'>bar</text>;
+    const resolvedJsx2 = <text className='foo'>bar</text>;
+
+    const rendered = renderToString(
+      <Suspense fallback={fallbackJsx1}>
+        <Suspender1>
+          <>
+            {resolvedJsx1}
+            <Suspense fallback={fallbackJsx2}>
+              <Suspender2>
+                <text className='foo' attr={Math.random()}>bar</text>
+              </Suspender2>
+            </Suspense>
+            {resolvedJsx2}
+          </>
+        </Suspender1>
+      </Suspense>,
+    );
+
+    expect(rendered.length).toBe(9);
+    expect(rendered[1].type).toStrictEqual(resolvedJsx1.type);
+    expect(rendered[4].type).toStrictEqual(fallbackJsx2.type);
+    expect(rendered[7].type).toStrictEqual(resolvedJsx2.type);
+  });
+
+  it('should render text with resolved suspense', async () => {
+    const { Suspender, suspended } = createSuspender();
+    suspended.resolve();
+    await waitSchedule();
+
+    const resolvedJsx = <text className='foo'>bar</text>;
+
+    const rendered = renderToString(
+      <Suspense fallback={<text>loading...</text>}>
+        <Suspender>
+          {resolvedJsx}
+        </Suspender>
+      </Suspense>,
+    );
+
+    expect(rendered.length).toBe(3);
+    expect(rendered[0]).toStrictEqual(0);
+    expect(rendered[1].type).toStrictEqual(resolvedJsx.type);
+    expect(rendered[2]).toStrictEqual(1);
+  });
+
+  it('should render text with nested suspense', async () => {
+    const { Suspender: Suspender1, suspended: suspended1 } = createSuspender();
+    const { Suspender: Suspender2, suspended: suspended2 } = createSuspender();
+    const { Suspender: Suspender3, suspended: suspended3 } = createSuspender();
+    suspended1.resolve();
+    suspended2.resolve();
+    suspended3.resolve();
+    await waitSchedule();
+
+    const resolvedJsx1 = <text className='foo'>bar1</text>;
+    const resolvedJsx2 = <text className='foo'>bar2</text>;
+    const resolvedJsx3 = <text className='foo'>bar3</text>;
+    const resolvedJsx4 = <text className='foo'>bar4</text>;
+    const resolvedJsx5 = <text className='foo'>bar5</text>;
+
+    const rendered = renderToString(
+      <view>
+        <Suspense fallback={<text>loading...</text>}>
+          <Suspender1>
+            {resolvedJsx1}
+            <Suspense fallback={<text>loading...</text>}>
+              <Suspender2>
+                {resolvedJsx2}
+                <Suspense fallback={<text>loading...</text>}>
+                  <Suspender3>
+                    {resolvedJsx3}
+                  </Suspender3>
+                </Suspense>
+                {resolvedJsx4}
+              </Suspender2>
+            </Suspense>
+            {resolvedJsx5}
+          </Suspender1>
+        </Suspense>
+      </view>,
+    );
+
+    expect(rendered.length).toBe(18);
+    expect(rendered[3].type).toStrictEqual(resolvedJsx1.type);
+    expect(rendered[6].type).toStrictEqual(resolvedJsx2.type);
+    expect(rendered[9].type).toStrictEqual(resolvedJsx3.type);
+    expect(rendered[12].type).toStrictEqual(resolvedJsx4.type);
+    expect(rendered[15].type).toStrictEqual(resolvedJsx5.type);
   });
 });
 
@@ -725,7 +936,6 @@ describe('renderOpcodesInto', () => {
       const componentVNodeC = vnodeC2.props.children;
       expect(componentVNodeC.type).toBe(Fragment);
       expect(componentVNodeC.props.children).toHaveLength(4);
-      expect(componentVNodeC.__c.constructor).toBe(Fragment);
       // FIXME(hzy): there is still a cycle reference
       expect(componentVNodeC.__c.__v).toBe(componentVNodeC);
       componentVNodeC.props.children.forEach((vnode) => {
@@ -762,8 +972,8 @@ describe('createElement', () => {
         {
           "children": undefined,
           "extraProps": undefined,
-          "id": -59,
-          "type": "__Card__:__snapshot_a94a8_test_44",
+          "id": -89,
+          "type": "__Card__:__snapshot_a94a8_test_74",
           "values": undefined,
         },
         1,
@@ -776,8 +986,8 @@ describe('createElement', () => {
         {
           "children": undefined,
           "extraProps": undefined,
-          "id": -60,
-          "type": "__Card__:__snapshot_a94a8_test_44",
+          "id": -90,
+          "type": "__Card__:__snapshot_a94a8_test_74",
           "values": undefined,
         },
         1,
@@ -798,8 +1008,8 @@ describe('createElement', () => {
         {
           "children": undefined,
           "extraProps": undefined,
-          "id": -61,
-          "type": "__Card__:__snapshot_a94a8_test_45",
+          "id": -91,
+          "type": "__Card__:__snapshot_a94a8_test_75",
           "values": undefined,
         },
         1,
@@ -812,8 +1022,8 @@ describe('createElement', () => {
         {
           "children": undefined,
           "extraProps": undefined,
-          "id": -62,
-          "type": "__Card__:__snapshot_a94a8_test_45",
+          "id": -92,
+          "type": "__Card__:__snapshot_a94a8_test_75",
           "values": undefined,
         },
         1,
@@ -874,7 +1084,9 @@ describe('createElement', () => {
 
 describe('cloneElement', () => {
   it('should clone component', () => {
-    function Comp() {}
+    function Comp() {
+    }
+
     const instance = <Comp prop1={1}>hello</Comp>;
     const clone = cloneElement(instance);
 
@@ -888,7 +1100,9 @@ describe('cloneElement', () => {
   });
 
   it('should merge new props', () => {
-    function Foo() {}
+    function Foo() {
+    }
+
     const instance = <Foo prop1={1} prop2={2} />;
     const clone = cloneElement(instance, { prop1: -1, newProp: -2 });
 
@@ -901,7 +1115,9 @@ describe('cloneElement', () => {
   });
 
   it('should override children if specified', () => {
-    function Foo() {}
+    function Foo() {
+    }
+
     const instance = <Foo>hello</Foo>;
     const clone = cloneElement(instance, null, 'world', '!');
 
@@ -911,7 +1127,9 @@ describe('cloneElement', () => {
   });
 
   it('should override children if null is provided as an argument', () => {
-    function Foo() {}
+    function Foo() {
+    }
+
     const instance = <Foo>hello</Foo>;
     const clone = cloneElement(instance, { children: 'bar' }, null);
 
@@ -921,7 +1139,9 @@ describe('cloneElement', () => {
   });
 
   it('should override key if specified', () => {
-    function Foo() {}
+    function Foo() {
+    }
+
     const instance = <Foo key='1'>hello</Foo>;
 
     let clone = cloneElement(instance);
@@ -934,9 +1154,15 @@ describe('cloneElement', () => {
   });
 
   it('should override ref if specified', () => {
-    function a() {}
-    function b() {}
-    function Foo() {}
+    function a() {
+    }
+
+    function b() {
+    }
+
+    function Foo() {
+    }
+
     const instance = <Foo ref={a}>hello</Foo>;
 
     let clone = cloneElement(instance);
@@ -954,6 +1180,7 @@ describe('cloneElement', () => {
         return <div style={{ color: props.color }}>thing</div>;
       }
     }
+
     Example.defaultProps = { color: 'blue' };
 
     const element = <Example color='red' />;