+++ packages/react/runtime/src/renderToOpcodes/index.ts
@@ -68,6 +68,7 @@ export function renderToString(vnode: any, context: any): any[] {
       undefined,
       parent,
       opcodes,
+      0,
     );
   } finally {
     // options._commit, we don't schedule any effects in this library right now,
@@ -99,7 +100,13 @@ export const __OpText = 3;
 function renderClassComponent(vnode, context) {
   const type = /** @type {import("preact").ComponentClass<typeof vnode.props>} */ (vnode.type);
 
-  const c = new type(vnode.props, context);
+  let c;
+  if (vnode[COMPONENT]) {
+    c = vnode[COMPONENT];
+    c.state = c[NEXT_STATE];
+  } else {
+    c = new type(vnode.props, context);
+  }
 
   vnode[COMPONENT] = c;
   c[VNODE] = vnode;
@@ -144,6 +151,7 @@ function _renderToString(
   selectValue,
   parent,
   opcodes,
+  opcodesLength,
 ) {
   // Ignore non-rendered VNodes/values
   if (vnode == null || vnode === true || vnode === false || vnode === '') {
@@ -165,7 +173,7 @@ function _renderToString(
       const child = vnode[i];
       if (child == null || typeof child === 'boolean') continue;
 
-      _renderToString(child, context, isSvgMode, selectValue, parent, opcodes);
+      _renderToString(child, context, isSvgMode, selectValue, parent, opcodes, opcodes.length);
     }
     return;
   }
@@ -241,11 +249,28 @@ function _renderToString(
     rendered = isTopLevelFragment ? rendered.props.children : rendered;
 
     // Recurse into children before invoking the after-diff hook
-    _renderToString(rendered, context, isSvgMode, selectValue, vnode, opcodes);
-    if (afterDiff) afterDiff(vnode);
-    vnode[PARENT] = undefined;
+    try {
+      _renderToString(rendered, context, isSvgMode, selectValue, vnode, opcodes, opcodes.length);
+    } catch (e) {
+      if (e && typeof e === 'object' && e.then && component && /* _childDidSuspend */ component.__c) {
+        component.setState({ /* _suspended */ __a: true });
 
-    if (ummountHook) ummountHook(vnode);
+        if (component[DIRTY]) {
+          rendered = renderClassComponent(vnode, context);
+          component = vnode[COMPONENT];
+
+          opcodes.length = opcodesLength;
+          _renderToString(rendered, context, isSvgMode, selectValue, vnode, opcodes, opcodes.length);
+        }
+      } else {
+        throw e;
+      }
+    } finally {
+      if (afterDiff) afterDiff(vnode);
+      vnode[PARENT] = undefined;
+
+      if (ummountHook) ummountHook(vnode);
+    }
 
     return;
   }
@@ -284,7 +309,7 @@ function _renderToString(
     opcodes.push(__OpText, children);
   } else if (children != null && children !== false && children !== true) {
     // recurse into this element VNode's children
-    _renderToString(children, context, false, selectValue, vnode, opcodes);
+    _renderToString(children, context, false, selectValue, vnode, opcodes, opcodes.length);
   }
 
   if (afterDiff) afterDiff(vnode);