+++ packages/react/testing-library/src/setupFiles/common/runtime-setup.js
@@ -35,9 +35,6 @@ globalThis.onInitWorkletRuntime = () => {
     onInitWorkletRuntime();
   }
 
-  if (process.env.DEBUG) {
-    console.log('initWorkletRuntime');
-  }
   lynx.setTimeout = setTimeout;
   lynx.setInterval = setInterval;
   lynx.clearTimeout = clearTimeout;
@@ -53,9 +50,6 @@ globalThis.onInitWorkletRuntime = () => {
 globalThis.onInjectMainThreadGlobals = (target) => {
   if (onInjectMainThreadGlobals) {
     onInjectMainThreadGlobals(target);
-  }
-  if (process.env.DEBUG) {
-    console.log('onInjectMainThreadGlobals');
   }
 
   snapshotInstanceManager.clear();
@@ -89,9 +83,6 @@ globalThis.onInjectMainThreadGlobals = (target) => {
 globalThis.onInjectBackgroundThreadGlobals = (target) => {
   if (onInjectBackgroundThreadGlobals) {
     onInjectBackgroundThreadGlobals(target);
-  }
-  if (process.env.DEBUG) {
-    console.log('onInjectBackgroundThreadGlobals');
   }
 
   backgroundSnapshotInstanceManager.clear();
@@ -137,9 +128,6 @@ globalThis.onResetLynxTestingEnv = () => {
   if (onResetLynxTestingEnv) {
     onResetLynxTestingEnv();
   }
-  if (process.env.DEBUG) {
-    console.log('onResetLynxTestingEnv');
-  }
 
   flushDelayedLifecycleEvents();
   destroyWorklet();
@@ -153,9 +141,6 @@ globalThis.onSwitchedToMainThread = () => {
   if (onSwitchedToMainThread) {
     onSwitchedToMainThread();
   }
-  if (process.env.DEBUG) {
-    console.log('onSwitchedToMainThread');
-  }
 
   setRoot(globalThis.__root);
   options.document = globalThis._document;
@@ -163,9 +148,6 @@ globalThis.onSwitchedToMainThread = () => {
 globalThis.onSwitchedToBackgroundThread = () => {
   if (onSwitchedToBackgroundThread) {
     onSwitchedToBackgroundThread();
-  }
-  if (process.env.DEBUG) {
-    console.log('onSwitchedToBackgroundThread');
   }
 
   setRoot(globalThis.__root);