+++ packages/third-party/tailwind-preset/src/types/index.d.ts
@@ -3,3 +3,4 @@
 // LICENSE file in the root directory of this source tree.
 export * from './tailwind-types.js';
 export * from './plugin-types.js';
+export * from './theme-types.js';